+++ tomcat/src/main/java/org/apache/coyote/response/HttpResponse.java
@@ -1,4 +1,6 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.response;
+
+import org.apache.coyote.response.ContentType;
 
 public class HttpResponse {
